RecipeSuggestionService/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
google-generativeai
pytest
pytest-asyncio
anyio